+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v0.0.2:
 
+* Extract the logstash patterns when installing logstash.
 * Upgrade to 1.1.10 version of logstash.
 * Prefer new notification syntax.
 * Ensure that patterns are correctly created with template resource.